==> 1_Sample.ipynb <==
import pandas as pd

from sklearn.model_selection import train_test_split

## 1. Sample

file_path = '/home/tomas/Documentos/DataSet/trabalho-apc-asn/yochi_base.xlsx'

df = pd.read_excel(file_path)


display(df)

y = df['BAD']
X = df.drop(axis=1, labels='BAD')

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20, random_state=59)

X_train, X_val, y_train, y_val = train_test_split(X_train, y_train, test_size=0.25, random_state=59)

print(f'X de treino: {X_train.shape}')
print(f'y de treino: {y_train.shape}')
print()

print(f'X de teste: {X_test.shape}')
print(f'y de teste: {y_test.shape}')
print()

print(f'X de validação: {X_val.shape}')
print(f'y de validação: {y_val.shape}')

file_path_1 = '/home/tomas/Documentos/gh-repos/Trabalho-APC-ASN/X_train.csv'
X_train.to_csv(file_path_1, index_label=False)

file_path_2 = '/home/tomas/Documentos/gh-repos/Trabalho-APC-ASN/y_train.csv'
y_train.to_csv(file_path_2, index_label=False)

file_path_3 ='/home/tomas/Documentos/gh-repos/Trabalho-APC-ASN/X_val.csv'
X_val.to_csv(file_path_3, index_label=False)

file_path_4 = '/home/tomas/Documentos/gh-repos/Trabalho-APC-ASN/y_val.csv'
y_val.to_csv(file_path_4, index_label=False)

file_path_5 = '/home/tomas/Documentos/gh-repos/Trabalho-APC-ASN/X_test.csv'
X_test.to_csv(file_path_5, index_label=False)

file_path_6 = '/home/tomas/Documentos/gh-repos/Trabalho-APC-ASN/y_test.csv'
y_test.to_csv(file_path_6, index_label=False)
