==> 4_Modelling.ipynb <==
import pandas as pd
import numpy as np

from sklearn.linear_model import LogisticRegression

from sklearn.metrics import make_scorer, precision_score, accuracy_score

from sklearn.model_selection import GridSearchCV

from sklearn.utils.class_weight import compute_class_weight


file_path_1 = '/home/tomas/Documentos/gh-repos/Trabalho-APC-ASN/X_train_M.csv'
train_t = pd.read_csv(file_path_1)

file_path_2 ='/home/tomas/Documentos/gh-repos/Trabalho-APC-ASN/X_val_M.csv'
val_t = pd.read_csv(file_path_2)

file_path_3 = '/home/tomas/Documentos/gh-repos/Trabalho-APC-ASN/X_test_M.csv'
test_t = pd.read_csv(file_path_3)

file_path_4 = '/home/tomas/Documentos/gh-repos/Trabalho-APC-ASN/y_train.csv'
y_train = np.ravel(pd.read_csv(file_path_4))

file_path_5 = '/home/tomas/Documentos/gh-repos/Trabalho-APC-ASN/y_val.csv'
y_val = np.ravel(pd.read_csv(file_path_5))

file_path_6 = '/home/tomas/Documentos/gh-repos/Trabalho-APC-ASN/y_test.csv'
y_test = np.ravel(pd.read_csv(file_path_6))

## Modelling

def class_weight_maker(y, classes):

  # Calcula a frequência das classs
  class_0_count = np.sum(y == 0)
  class_1_count = np.sum(y == 1)

  # Calcula o número total observações
  total_samples = len(y)

  # Calculate class weights using "balanced" mode
  class_weights = compute_class_weight('balanced', classes=classes, y=y)
  class_weight = {0: class_weights[0], 1: class_weights[1]}
  return class_weight

weight = class_weight_maker(y_train, [0, 1])
weight

modelo = LogisticRegression(tol=0.0001,
                            random_state=59, 
                            max_iter=10000,
                            #class_weight=weight,
                            )
#
#
#
#
param_grid = {'penalty': ['l1', 'l2'],
              'C': np.logspace(-1, 1, 20),
              'solver': ['liblinear', 'saga'],
              }

grid_search = GridSearchCV(modelo, param_grid, cv=10, n_jobs=-1, scoring='accuracy')
grid_search.fit(train_t, y_train)

print(grid_search.best_params_)


from sklearn.metrics import accuracy_score, classification_report

best_params = grid_search.best_params_
best_model = grid_search.best_estimator_


# Predict using the best model
y_pred = best_model.predict(val_t)

# Calculate accuracy and print classification report
precision = precision_score(y_val, y_pred)
print("Precision:", precision)
print("Best hyperparameters:", best_params)
print("Classification Report:\n", classification_report(y_val, y_pred))


from sklearn.metrics import accuracy_score, classification_report

best_params = grid_search.best_params_
best_model = grid_search.best_estimator_


# Predict using the best model
y_pred = best_model.predict(val_t)

# Calculate accuracy and print classification report
accuracy = accuracy_score(y_val, y_pred)
print("Accuracy:", accuracy)
print("Best hyperparameters:", best_params)
print("Classification Report:\n", classification_report(y_test, y_pred))


hiperparametros = {'penalty':'l1',
                   'tol':0.0001,
                   'C':0.0048, 
                   'random_state':59, 
                   'solver':'liblinear',
                   'max_iter':10000,
                   'class_weight':weight}
